# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /masters/warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      summary: List warehouses
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: city_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Warehouse list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warehouse"
    post:
      operationId: createWarehouse
      tags: [warehouses]
      summary: Create warehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "201":
          description: Warehouse created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /masters/warehouses/{id}:
    put:
      operationId: updateWarehouse
      tags: [warehouses]
      summary: Update warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WarehouseInput"
      responses:
        "200":
          description: Warehouse updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warehouse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      summary: Delete warehouse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Warehouse deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Cannot delete warehouse in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Warehouse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cityId:
          type: integer
        cityName:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        createdAt:
          type: string
          nullable: true
      required: [id, name, cityId]
    WarehouseInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        cityId:
          type: integer
        address:
          type: string
          nullable: true
      required: [name, cityId]
...