# lib/api-zod/src/generated/types/warehouse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface Warehouse {
  id: number;
  name: string;
  cityId: number;
  cityName?: string | null;
  address?: string | null;
  createdAt?: string | null;
}