# lib/api-zod/src/generated/types/warehouseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface WarehouseInput {
  /** @minLength 1 */
  name: string;
  cityId: number;
  address?: string | null;
}